Refactored into go packages
* Created `view` package. Views are structs that we can navigate between
* Created `cfg` package. A top-level struct that holds configuration of the program
* Created `style` package. Mostly global vars that defines lipgloss styles that we can apply to our views
