+++ cmd/main.go
@@ -5,19 +5,15 @@ import (
 
 	"errors"
 	"fmt"
-	"io"
 	"os"
 	"path"
 	"path/filepath"
-	"strconv"
-	"strings"
-	"time"
-
-	"k8s.io/apimachinery/pkg/util/duration"
 
 	"k8s.io/client-go/tools/clientcmd"
-	"k8s.io/client-go/tools/clientcmd/api"
 
+	"github.com/amimof/kubecfg/pkg/cfg"
+	"github.com/amimof/kubecfg/pkg/style"
+	"github.com/amimof/kubecfg/pkg/view"
 	"github.com/charmbracelet/bubbles/list"
 	tea "github.com/charmbracelet/bubbletea"
 	"github.com/charmbracelet/lipgloss"
@@ -45,112 +41,24 @@ var (
 
 	// Errors
 	ErrExist = errors.New("file already exists")
-
-	// Colors
-	subtle = lipgloss.AdaptiveColor{Light: "#D9DCCF", Dark: "#383838"}
-	body   = lipgloss.AdaptiveColor{Light: "#343433", Dark: "#C1C6B2"}
-
-	// The list pane that displays the files on the left side
-	listPaneWidth  = 60
-	listPaneHeight = 20
-	listPane       = lipgloss.NewStyle().
-			Width(listPaneWidth).
-			Height(listPaneHeight).
-			PaddingTop(1)
-
-	errorStyle = lipgloss.NewStyle().
-			Background(lipgloss.Color("#d75f00")).
-			Foreground(lipgloss.Color("#eeeeee")).
-			MarginRight(2)
-
-	// The pane on the right that displays details about the currently selected file
-	detailsPane = lipgloss.NewStyle().
-			BorderStyle(lipgloss.RoundedBorder()).
-			BorderForeground(lipgloss.Color("#7D56F4")).
-			Foreground(lipgloss.Color("#FAFAFA")).
-			PaddingLeft(1).
-			PaddingRight(1).
-			Margin(0).
-			Align(lipgloss.Left).
-			Width(24).
-			Height(20)
-
-	// This is a header usually display inside of the details pane
-	detailsHeader = lipgloss.NewStyle().
-			Foreground(subtle).
-			MarginRight(2)
-
-	// This is text displayed containing actual values. Usually used together with detailsHeader
-	detailsContent = lipgloss.NewStyle().
-			Foreground(body).
-			PaddingBottom(1)
 )
 
 type state int
 
 // Top level app model
 type model struct {
-	state        state
-	warningModel warningModel
-	defaultModel defaultModel
+	state            state
+	backupConfigView view.BackupConfigView
+	mainView         view.MainView
 }
 
-type defaultModel struct {
-	items    []*Item
-	list     list.Model
-	selected int
-	errMsg   string
-}
-
-type Item struct {
-	config     *api.Config
-	info       os.FileInfo
-	title      string
-	desc       string
-	isSelected bool
-}
-
-func (i Item) Title() string {
-	s := i.title
-	if i.isSelected {
-		s = lipgloss.NewStyle().Foreground(lipgloss.Color("#87ffaf")).Render(i.title + " •")
-	}
-	return s
-}
-func (i Item) Description() string {
-	return i.desc
-}
-func (i Item) FilterValue() string {
-	return i.title
-}
-
-func newDefaultModel() defaultModel {
-	return defaultModel{}
-}
-
-func New() *model {
+func New(c *cfg.Cfg) *model {
 	r := &model{
-		state:        stateShowDefault,
-		defaultModel: newDefaultModel(),
-		warningModel: newWarningModel(),
+		state:            stateShowDefault,
+		backupConfigView: view.NewBackupConfigView(c),
+		mainView:         view.NewMainView(c),
 	}
 	return r
-}
-
-func (r *defaultModel) Add(i *Item) {
-	r.items = append(r.items, i)
-}
-
-func (r *defaultModel) List() []list.Item {
-	var items []list.Item
-	for _, i := range r.items {
-		items = append(items, i)
-	}
-	return items
-}
-
-func (r defaultModel) Init() tea.Cmd {
-	return nil
 }
 
 func (r *model) Init() tea.Cmd {
@@ -164,191 +72,45 @@ func (r *model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 	switch msg := msg.(type) {
 	case tea.WindowSizeMsg:
 		h, v := lipgloss.NewStyle().Margin(1, 2).GetFrameSize()
-		r.defaultModel.list.SetSize(msg.Width-h, msg.Height-v)
-		w := msg.Width - listPaneWidth - h
+		r.mainView.List.SetSize(msg.Width-h, msg.Height-v)
+		w := msg.Width - style.ListPaneWidth - h
 		he := msg.Height - v
-		detailsPane = detailsPane.Width(w)
-		detailsPane = detailsPane.Height(he - 1)
-	case OverwriteConfigMsg:
+		style.DetailsPane = style.DetailsPane.Width(w)
+		style.DetailsPane = style.DetailsPane.Height(he - 1)
+	case view.ChangeViewMsg:
 		r.state = stateShowDefault
-		newDetaultModel, cmd := r.defaultModel.update(msg)
-		r.defaultModel = newDetaultModel
+		newDetaultModel, cmd := r.mainView.UpdateView(msg)
+		r.mainView = newDetaultModel
 		cmds = append(cmds, cmd)
 
 	}
 
 	switch r.state {
 	case stateShowDefault:
-		newDetaultModel, cmd := r.defaultModel.update(msg)
-		r.defaultModel = newDetaultModel
+		newDetaultModel, cmd := r.mainView.UpdateView(msg)
+		r.mainView = newDetaultModel
 		cmds = append(cmds, cmd)
 	case stateShowWarning:
-		newWarningModel, cmd := r.warningModel.update(msg)
-		r.warningModel = newWarningModel
+		newbackupConfigView, cmd := r.backupConfigView.UpdateView(msg)
+		r.backupConfigView = newbackupConfigView
 		cmds = append(cmds, cmd)
 	}
 
 	return r, tea.Batch(cmds...)
-}
-
-func (r defaultModel) update(msg tea.Msg) (defaultModel, tea.Cmd) {
-	var cmds []tea.Cmd
-	switch msg := msg.(type) {
-	case tea.KeyMsg:
-		switch keypress := msg.String(); keypress {
-		case "ctrl+c":
-			return r, tea.Quit
-		case "down":
-			if r.selected < len(r.items)-1 {
-				r.selected++
-			}
-		case "up":
-			if r.selected > 0 {
-				r.selected--
-			}
-		case "enter":
-			p, err := parseArgs()
-			if err != nil {
-				r.errMsg = err.Error()
-			}
-			err = os.Remove(path.Join(p, "config"))
-			if err != nil {
-				r.errMsg = err.Error()
-			}
-			err = os.Symlink(path.Join(p, r.items[r.selected].info.Name()), path.Join(p, "config"))
-			if err != nil {
-				r.errMsg = err.Error()
-			}
-			r.selectItem(r.selected)
-		}
-	case tea.WindowSizeMsg:
-		h, v := lipgloss.NewStyle().Margin(1, 2).GetFrameSize()
-		r.list.SetSize(msg.Width-h, msg.Height-v)
-		w := msg.Width - listPaneWidth - h
-		he := msg.Height - v
-		detailsPane = detailsPane.Width(w)
-		detailsPane = detailsPane.Height(he - 1)
-	}
-	l, cmd := r.list.Update(msg)
-	r.list = l
-	cmds = append(cmds, cmd)
-	return r, tea.Batch(cmds...)
-}
-
-func (r *defaultModel) selectItem(i int) {
-	for _, i := range r.items {
-		i.isSelected = false
-	}
-	r.items[i].isSelected = true
-}
-
-func (r defaultModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
-	return r.update(msg)
-}
-
-func (r defaultModel) View() string {
-	curContext := lipgloss.JoinVertical(
-		lipgloss.Top,
-		detailsHeader.Render("Current Context"),
-		detailsContent.Render(r.getCurrentContext()),
-	)
-	contextCount := lipgloss.JoinVertical(
-		lipgloss.Left,
-		detailsHeader.Render("Contexts"),
-		detailsContent.Render(r.getContextCount()),
-	)
-	clusterCount := lipgloss.JoinVertical(
-		lipgloss.Left,
-		detailsHeader.Render("Clusters"),
-		detailsContent.Render(r.getClusterCount()),
-	)
-	userCount := lipgloss.JoinVertical(
-		lipgloss.Left,
-		detailsHeader.Render("Users"),
-		detailsContent.Render(r.getUserCount()),
-	)
-	counts := lipgloss.JoinHorizontal(
-		lipgloss.Left,
-		contextCount,
-		clusterCount,
-		userCount,
-	)
-	modified := lipgloss.JoinVertical(
-		lipgloss.Left,
-		detailsHeader.Render("Last Modified"),
-		detailsContent.Render(r.getLastModified()),
-	)
-	sizeBytes := lipgloss.JoinVertical(
-		lipgloss.Left,
-		detailsHeader.Render("Size"),
-		detailsContent.Render(r.getSize()),
-	)
-
-	details := lipgloss.JoinVertical(
-		lipgloss.Top,
-		counts,
-		curContext,
-		modified,
-		sizeBytes,
-		errorStyle.Render(r.errMsg),
-	)
-	return lipgloss.JoinHorizontal(lipgloss.Top, listPane.Render(r.list.View()), detailsPane.Render(details))
 }
 
 func (r *model) View() string {
 	var s string
 	switch r.state {
 	case stateShowWarning:
-		return r.warningModel.View()
+		return r.backupConfigView.View()
 	case stateShowDefault:
-		return r.defaultModel.View()
+		return r.mainView.View()
 	}
 
 	return s
 
 }
-
-func (r *defaultModel) getContextCount() string {
-	str := strconv.Itoa(len(r.items[r.selected].config.Contexts))
-	return str
-}
-
-func (r *defaultModel) getClusterCount() string {
-	str := strconv.Itoa(len(r.items[r.selected].config.Clusters))
-	return str
-}
-
-func (r *defaultModel) getUserCount() string {
-	str := strconv.Itoa(len(r.items[r.selected].config.AuthInfos))
-	return str
-}
-
-func (r *defaultModel) getCurrentContext() string {
-	return r.items[r.selected].config.CurrentContext
-}
-
-func (r *defaultModel) getLastModified() string {
-	return duration.HumanDuration(time.Since(r.items[r.selected].info.ModTime())) + " ago"
-}
-
-func ByteCountSI(b int64) string {
-	const unit = 1000
-	if b < unit {
-		return fmt.Sprintf("%d B", b)
-	}
-	div, exp := int64(unit), 0
-	for n := b / unit; n >= unit; n /= unit {
-		div *= unit
-		exp++
-	}
-	return fmt.Sprintf("%.1f %cB",
-		float64(b)/float64(div), "kMGTPE"[exp])
-}
-
-func (r *defaultModel) getSize() string {
-	return ByteCountSI(int64(r.items[r.selected].info.Size()))
-}
-
 func usage() {
 	fmt.Fprint(os.Stderr, "Usage:\n")
 	fmt.Fprint(os.Stderr, "  kubecfg [PATH] <flags>\n\n")
@@ -392,7 +154,9 @@ func main() {
 		return
 	}
 
-	res := New()
+	res := New(&cfg.Cfg{
+		Path: p,
+	})
 
 	// Evaluate if symlink ~/.kube/config already exists since we don't want to overwrite the users config file
 	if err = CheckConfig(path.Join(p, "config")); errors.Is(err, ErrExist) {
@@ -408,18 +172,13 @@ func main() {
 			if err != nil {
 				return nil
 			}
-			i := &Item{
-				config: config,
-				info:   info,
-				title:  info.Name(),
-				desc:   fmt.Sprintf("%d contexts", len(config.Contexts)),
-			}
+			i := res.mainView.NewItem(config, info, info.Name(), fmt.Sprintf("%d contexts", len(config.Contexts)))
 			if pa, err := filepath.EvalSymlinks(path.Join(p, "config")); err == nil {
 				if pa == filePath {
-					i.isSelected = true
+					i.IsSelected = true
 				}
 			}
-			res.defaultModel.Add(i)
+			res.mainView.Add(i)
 		}
 		return nil
 	})
@@ -428,8 +187,8 @@ func main() {
 		os.Exit(1)
 	}
 
-	res.defaultModel.list = list.New(res.defaultModel.List(), list.NewDefaultDelegate(), 0, 0)
-	res.defaultModel.list.Title = p
+	res.mainView.List = list.New(res.mainView.ListItems(), list.NewDefaultDelegate(), 0, 0)
+	res.mainView.List.Title = p
 
 	prog := tea.NewProgram(res, tea.WithAltScreen())
 	if _, err := prog.Run(); err != nil {
@@ -452,159 +211,4 @@ func CheckConfig(p string) error {
 		return ErrExist
 	}
 	return nil
-}
-
-var (
-	choices = []string{
-		"Make a backup and create a symlink to it",
-		"Delete it. I want a fresh start",
-		"Do nothing, I'm too paranoid to know what to do",
-	}
-	titleStyle = lipgloss.NewStyle().
-			Foreground(lipgloss.Color("#7D56F4")).
-			Margin(1, 2).
-			Width(70)
-	selectedStyle = lipgloss.NewStyle().
-			Foreground(lipgloss.Color("230")).
-			MarginLeft(2)
-	inactiveStyle = lipgloss.NewStyle().
-			Foreground(lipgloss.Color("212")).
-			MarginLeft(2)
-)
-
-type warningModel struct {
-	cursor int
-	choice string
-	errMsg string
-}
-
-func newWarningModel() warningModel {
-	return warningModel{}
-}
-
-func (m warningModel) Init() tea.Cmd {
-	return nil
-}
-
-type OverwriteConfigMsg string
-
-func OverwriteConfigCmd() tea.Cmd {
-	return func() tea.Msg {
-		return OverwriteConfigMsg("Overwriting config")
-	}
-}
-
-func (m warningModel) update(msg tea.Msg) (warningModel, tea.Cmd) {
-	var cmds []tea.Cmd
-	switch msg := msg.(type) {
-	case tea.KeyMsg:
-		switch msg.String() {
-		case "ctrl+c", "q", "esc":
-			return m, tea.Quit
-		case "enter":
-			m.choice = choices[m.cursor]
-			err := m.executeChoice(m.cursor)
-			if err != nil {
-				m.errMsg = err.Error()
-			} else {
-				cmds = append(cmds, OverwriteConfigCmd())
-			}
-
-		case "down", "j":
-			m.cursor++
-			if m.cursor >= len(choices) {
-				m.cursor = 0
-			}
-		case "up", "k":
-			m.cursor--
-			if m.cursor < 0 {
-				m.cursor = len(choices) - 1
-			}
-		}
-	}
-	return m, tea.Batch(cmds...)
-}
-
-func (m *warningModel) executeChoice(i int) error {
-	p, err := parseArgs()
-	if err != nil {
-		return err
-	}
-	switch i {
-	case 0:
-		err := m.copyFile(path.Join(p, "config"), path.Join(p, fmt.Sprintf("%s_kubecfg-backup", "config")))
-		if err != nil {
-			return err
-		}
-		err = os.Remove(path.Join(p, "config"))
-		if err != nil {
-			return err
-		}
-		err = m.createLink(fmt.Sprintf("%s_kubecfg-backup", "config"), path.Join(p, "config"))
-		if err != nil {
-			return err
-		}
-	case 1:
-	case 2:
-	default:
-	}
-	return nil
-}
-
-func (m *warningModel) copyFile(src, dst string) error {
-	// srcStat, err := os.Stat(src)
-	// if err != nil {
-	// 	return err
-	// }
-	srcFile, err := os.Open(src)
-	if err != nil {
-		return err
-	}
-	defer srcFile.Close()
-
-	dstFile, err := os.Create(dst)
-	if err != nil {
-		return err
-	}
-	defer dstFile.Close()
-
-	_, err = io.Copy(dstFile, srcFile)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (m *warningModel) createLink(target, link string) error {
-	return os.Symlink(target, link)
-}
-
-// func (m *warningModel) SwitchConfig(src, dst string) error {
-// 	if err := m.copyFile(src, dst); err != nil {
-// 		return err
-// 	}
-// 	if err :=
-// }
-
-func (m warningModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
-	return m.update(msg)
-}
-
-func (m warningModel) View() string {
-	s := strings.Builder{}
-	s.WriteString(titleStyle.Render("Kubecfg uses symbolic links to switch between different kubeconfig files. However, an existing kubeconfig file was found. What do you want me to do with it?") + "\n")
-	for i := 0; i < len(choices); i++ {
-		if m.cursor == i {
-			s.WriteString(selectedStyle.Render("➜ " + choices[i]))
-		} else {
-			s.WriteString(inactiveStyle.Render("  " + choices[i]))
-		}
-		s.WriteString("\n")
-	}
-
-	if len(m.errMsg) > 0 {
-		s.WriteString("\n  " + errorStyle.Render(fmt.Sprintf("Error: %s", m.errMsg)))
-	}
-
-	return s.String()
 }
