+++ pkg/cfg/cfg.go
@@ -1,0 +1,5 @@
+package cfg
+
+type Cfg struct {
+	Path string
+}
